backend/requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
amqp==5.0.6
anyio==3.3.3
APScheduler==3.8.0
asgiref==3.4.1
async-timeout==3.0.1
asyncpg==0.24.0
autopep8==1.5.7
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.1.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
databases==0.5.2
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.68.1
fastapi-mail==1.0.1
fastapi-utils==0.2.1
flower==1.0.0
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.19.0
humanize==3.12.0
idna==3.2
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
packaging==21.0
passlib==1.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.25
starlette==0.14.2
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
yapf==0.31.0
